separate file for application-related code
