+++ kwm/application.cpp
@@ -1,0 +1,98 @@
+#include "application.h"
+#include "window.h"
+
+extern kwm_tiling KWMTiling;
+extern kwm_cache KWMCache;
+
+void AllowRoleForApplication(std::string Application, std::string Role)
+{
+    std::map<std::string, std::vector<CFTypeRef> >::iterator It = KWMTiling.AllowedWindowRoles.find(Application);
+    if(It == KWMTiling.AllowedWindowRoles.end())
+        KWMTiling.AllowedWindowRoles[Application] = std::vector<CFTypeRef>();
+
+    CFStringRef RoleRef = CFStringCreateWithCString(NULL, Role.c_str(), kCFStringEncodingMacRoman);
+    KWMTiling.AllowedWindowRoles[Application].push_back(RoleRef);
+}
+
+bool IsAppSpecificWindowRole(window_info *Window, CFTypeRef Role, CFTypeRef SubRole)
+{
+    std::map<std::string, std::vector<CFTypeRef> >::iterator It = KWMTiling.AllowedWindowRoles.find(Window->Owner);
+    if(It != KWMTiling.AllowedWindowRoles.end())
+    {
+        std::vector<CFTypeRef> &WindowRoles = It->second;
+        for(std::size_t RoleIndex = 0; RoleIndex < WindowRoles.size(); ++RoleIndex)
+        {
+            if(CFEqual(Role, WindowRoles[RoleIndex]) || CFEqual(SubRole, WindowRoles[RoleIndex]))
+                return true;
+        }
+    }
+
+    return false;
+}
+
+bool IsApplicationFloating(window_info *Window)
+{
+    for(std::size_t WindowIndex = 0; WindowIndex < KWMTiling.FloatingAppLst.size(); ++WindowIndex)
+    {
+        if(Window->Owner == KWMTiling.FloatingAppLst[WindowIndex])
+            return true;
+    }
+
+    return false;
+}
+
+bool IsApplicationCapturedByScreen(window_info *Window)
+{
+    return KWMTiling.CapturedAppLst.find(Window->Owner) != KWMTiling.CapturedAppLst.end();
+}
+
+bool IsApplicationInCache(int PID, std::vector<AXUIElementRef> *Elements)
+{
+    bool Result = false;
+    std::map<int, std::vector<AXUIElementRef> >::iterator It = KWMCache.WindowRefs.find(PID);
+
+    if(It != KWMCache.WindowRefs.end())
+    {
+        *Elements = It->second;
+        Result = true;
+    }
+
+    return Result;
+}
+
+bool GetWindowRefFromCache(window_info *Window, AXUIElementRef *WindowRef)
+{
+    std::vector<AXUIElementRef> Elements;
+    bool IsCached = IsApplicationInCache(Window->PID, &Elements);
+
+    if(IsCached)
+    {
+        for(std::size_t ElementIndex = 0; ElementIndex < Elements.size(); ++ElementIndex)
+        {
+            if(GetWindowIDFromRef(Elements[ElementIndex]) == Window->WID)
+            {
+                *WindowRef = Elements[ElementIndex];
+                return true;
+            }
+        }
+    }
+
+    if(!IsCached)
+        KWMCache.WindowRefs[Window->PID] = std::vector<AXUIElementRef>();
+
+    return false;
+}
+
+void FreeWindowRefCache(int PID)
+{
+    std::map<int, std::vector<AXUIElementRef> >::iterator It = KWMCache.WindowRefs.find(PID);
+
+    if(It != KWMCache.WindowRefs.end())
+    {
+        int NumElements = KWMCache.WindowRefs[PID].size();
+        for(int RefIndex = 0; RefIndex < NumElements; ++RefIndex)
+            CFRelease(KWMCache.WindowRefs[PID][RefIndex]);
+
+        KWMCache.WindowRefs[PID].clear();
+    }
+}
