+++ kwm/display.cpp
@@ -2,6 +2,7 @@
 #include "space.h"
 #include "container.h"
 #include "tree.h"
+#include "application.h"
 #include "window.h"
 #include "helpers.h"
