+++ kwm/window.cpp
@@ -4,6 +4,7 @@
 #include "display.h"
 #include "space.h"
 #include "tree.h"
+#include "application.h"
 #include "notifications.h"
 #include "border.h"
 #include "helpers.h"
@@ -28,32 +29,6 @@ bool WindowsAreEqual(window_info *Window, window_info *Match)
            Window->WID == Match->WID &&
            Window->Layer == Match->Layer)
             return true;
-    }
-
-    return false;
-}
-
-void AllowRoleForApplication(std::string Application, std::string Role)
-{
-    std::map<std::string, std::vector<CFTypeRef> >::iterator It = KWMTiling.AllowedWindowRoles.find(Application);
-    if(It == KWMTiling.AllowedWindowRoles.end())
-        KWMTiling.AllowedWindowRoles[Application] = std::vector<CFTypeRef>();
-
-    CFStringRef RoleRef = CFStringCreateWithCString(NULL, Role.c_str(), kCFStringEncodingMacRoman);
-    KWMTiling.AllowedWindowRoles[Application].push_back(RoleRef);
-}
-
-bool IsAppSpecificWindowRole(window_info *Window, CFTypeRef Role, CFTypeRef SubRole)
-{
-    std::map<std::string, std::vector<CFTypeRef> >::iterator It = KWMTiling.AllowedWindowRoles.find(Window->Owner);
-    if(It != KWMTiling.AllowedWindowRoles.end())
-    {
-        std::vector<CFTypeRef> &WindowRoles = It->second;
-        for(std::size_t RoleIndex = 0; RoleIndex < WindowRoles.size(); ++RoleIndex)
-        {
-            if(CFEqual(Role, WindowRoles[RoleIndex]) || CFEqual(SubRole, WindowRoles[RoleIndex]))
-                return true;
-        }
     }
 
     return false;
@@ -133,22 +108,6 @@ bool FilterWindowList(screen_info *Screen)
 
     KWMTiling.WindowLst = FilteredWindowLst;
     return true;
-}
-
-bool IsApplicationCapturedByScreen(window_info *Window)
-{
-    return KWMTiling.CapturedAppLst.find(Window->Owner) != KWMTiling.CapturedAppLst.end();
-}
-
-bool IsApplicationFloating(window_info *Window)
-{
-    for(std::size_t WindowIndex = 0; WindowIndex < KWMTiling.FloatingAppLst.size(); ++WindowIndex)
-    {
-        if(Window->Owner == KWMTiling.FloatingAppLst[WindowIndex])
-            return true;
-    }
-
-    return false;
 }
 
 bool IsFocusedWindowFloating()
@@ -1911,57 +1870,6 @@ bool GetWindowRef(window_info *Window, AXUIElementRef *WindowRef)
 
     CFRelease(App);
     return Found;
-}
-
-bool IsApplicationInCache(int PID, std::vector<AXUIElementRef> *Elements)
-{
-    bool Result = false;
-    std::map<int, std::vector<AXUIElementRef> >::iterator It = KWMCache.WindowRefs.find(PID);
-
-    if(It != KWMCache.WindowRefs.end())
-    {
-        *Elements = It->second;
-        Result = true;
-    }
-
-    return Result;
-}
-
-bool GetWindowRefFromCache(window_info *Window, AXUIElementRef *WindowRef)
-{
-    std::vector<AXUIElementRef> Elements;
-    bool IsCached = IsApplicationInCache(Window->PID, &Elements);
-
-    if(IsCached)
-    {
-        for(std::size_t ElementIndex = 0; ElementIndex < Elements.size(); ++ElementIndex)
-        {
-            if(GetWindowIDFromRef(Elements[ElementIndex]) == Window->WID)
-            {
-                *WindowRef = Elements[ElementIndex];
-                return true;
-            }
-        }
-    }
-
-    if(!IsCached)
-        KWMCache.WindowRefs[Window->PID] = std::vector<AXUIElementRef>();
-
-    return false;
-}
-
-void FreeWindowRefCache(int PID)
-{
-    std::map<int, std::vector<AXUIElementRef> >::iterator It = KWMCache.WindowRefs.find(PID);
-
-    if(It != KWMCache.WindowRefs.end())
-    {
-        int NumElements = KWMCache.WindowRefs[PID].size();
-        for(int RefIndex = 0; RefIndex < NumElements; ++RefIndex)
-            CFRelease(KWMCache.WindowRefs[PID][RefIndex]);
-
-        KWMCache.WindowRefs[PID].clear();
-    }
 }
 
 void GetWindowInfo(const void *Key, const void *Value, void *Context)
